Move entity_service.py to common/
